feature(population): Create a map to guarantee users and organizations order
Since there are some issues with arrays of references across platforms, users are mixed when populating if they are not stored in mongo natural order. This made the library guarantee the order of them. It increases the number of operations done but the overall complexity stays O(n).

Hopefully, no breaking changes